Rename module `stellar.stars` to `stellar.stellar` to avoid import issue
