+++ scopesim_templates/__init__.py
@@ -15,7 +15,7 @@ from . import metis
 #               "Always use this syntax: from module.submodule import function",
 #               DeprecationWarning, stacklevel=2)
 from .misc.misc import source_from_image, source_from_file, source_from_cube
-from .stellar.stars import star, stars, star_grid, star_field
+from .stellar.stellar import star, stars, star_grid, star_field
 from .stellar.clusters import cluster
 from .extragalactic.galaxies import galaxy, galaxy3d, elliptical
 from .calibration.calibration import empty_sky
