+++ scopesim_templates/stellar/stellar.py
